Listening to jslint
